@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: ASTQueryProvider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_interfaces -destination test/mocks/app/ast_query/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces ASTQueryProvider
+//
 // Package mock_interfaces is a generated GoMock package.
 package mock_interfaces
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	soap "github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockASTQueryProvider is a mock of ASTQueryProvider interface.
@@ -59,7 +63,7 @@ func (m *MockASTQueryProvider) GetQueryID(arg0, arg1, arg2, arg3 string) (string
 }
 
 // GetQueryID indicates an expected call of GetQueryID.
-func (mr *MockASTQueryProviderMockRecorder) GetQueryID(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+func (mr *MockASTQueryProviderMockRecorder) GetQueryID(arg0, arg1, arg2, arg3 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetQueryID", reflect.TypeOf((*MockASTQueryProvider)(nil).GetQueryID), arg0, arg1, arg2, arg3)
 }
