@@ -1,13 +1,17 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: ASTQueryIDProvider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_app_ast_query_id -destination test/mocks/app/ast_query_id/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces ASTQueryIDProvider
+//
 // Package mock_app_ast_query_id is a generated GoMock package.
 package mock_app_ast_query_id
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockASTQueryIDProvider is a mock of ASTQueryIDProvider interface.
@@ -43,7 +47,7 @@ func (m *MockASTQueryIDProvider) GetQueryID(arg0, arg1, arg2, arg3 string) (stri
 }
 
 // GetQueryID indicates an expected call of GetQueryID.
-func (mr *MockASTQueryIDProviderMockRecorder) GetQueryID(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+func (mr *MockASTQueryIDProviderMockRecorder) GetQueryID(arg0, arg1, arg2, arg3 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetQueryID", reflect.TypeOf((*MockASTQueryIDProvider)(nil).GetQueryID), arg0, arg1, arg2, arg3)
 }
