@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: QueryMappingRepo)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_interfaces -destination test/mocks/app/ast_query_mapping/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces QueryMappingRepo
+//
 // Package mock_interfaces is a generated GoMock package.
 package mock_interfaces
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	querymapping "github.com/checkmarxDev/ast-sast-export/internal/app/querymapping"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockQueryMappingRepo is a mock of QueryMappingRepo interface.
@@ -43,7 +47,7 @@ func (m *MockQueryMappingRepo) AddQueryMapping(arg0, arg1, arg2, arg3 string) er
 }
 
 // AddQueryMapping indicates an expected call of AddQueryMapping.
-func (mr *MockQueryMappingRepoMockRecorder) AddQueryMapping(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
+func (mr *MockQueryMappingRepoMockRecorder) AddQueryMapping(arg0, arg1, arg2, arg3 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddQueryMapping", reflect.TypeOf((*MockQueryMappingRepo)(nil).AddQueryMapping), arg0, arg1, arg2, arg3)
 }
