@@ -1,13 +1,17 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/export (interfaces: Exporter)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_app_export -destination test/mocks/app/export/mock_exporter.go github.com/checkmarxDev/ast-sast-export/internal/app/export Exporter
+//
 // Package mock_app_export is a generated GoMock package.
 package mock_app_export
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockExporter is a mock of Exporter interface.
@@ -42,7 +46,7 @@ func (m *MockExporter) AddFile(arg0 string, arg1 []byte) error {
 }
 
 // AddFile indicates an expected call of AddFile.
-func (mr *MockExporterMockRecorder) AddFile(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockExporterMockRecorder) AddFile(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddFile", reflect.TypeOf((*MockExporter)(nil).AddFile), arg0, arg1)
 }
@@ -56,7 +60,7 @@ func (m *MockExporter) AddFileWithDataSource(arg0 string, arg1 func() ([]byte, e
 }
 
 // AddFileWithDataSource indicates an expected call of AddFileWithDataSource.
-func (mr *MockExporterMockRecorder) AddFileWithDataSource(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockExporterMockRecorder) AddFileWithDataSource(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddFileWithDataSource", reflect.TypeOf((*MockExporter)(nil).AddFileWithDataSource), arg0, arg1)
 }
@@ -84,7 +88,7 @@ func (m *MockExporter) CreateDir(arg0 string) error {
 }
 
 // CreateDir indicates an expected call of CreateDir.
-func (mr *MockExporterMockRecorder) CreateDir(arg0 interface{}) *gomock.Call {
+func (mr *MockExporterMockRecorder) CreateDir(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateDir", reflect.TypeOf((*MockExporter)(nil).CreateDir), arg0)
 }
@@ -100,7 +104,7 @@ func (m *MockExporter) CreateExportPackage(arg0, arg1 string) (string, string, e
 }
 
 // CreateExportPackage indicates an expected call of CreateExportPackage.
-func (mr *MockExporterMockRecorder) CreateExportPackage(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockExporterMockRecorder) CreateExportPackage(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateExportPackage", reflect.TypeOf((*MockExporter)(nil).CreateExportPackage), arg0, arg1)
 }
