@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/metadata (interfaces: Provider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_app_metadata -destination test/mocks/app/metadata/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/metadata Provider
+//
 // Package mock_app_metadata is a generated GoMock package.
 package mock_app_metadata
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	metadata "github.com/checkmarxDev/ast-sast-export/internal/app/metadata"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockProvider is a mock of Provider interface.
@@ -44,7 +48,7 @@ func (m *MockProvider) GetMetadataRecord(arg0 string, arg1 []*metadata.Query) (*
 }
 
 // GetMetadataRecord indicates an expected call of GetMetadataRecord.
-func (mr *MockProviderMockRecorder) GetMetadataRecord(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockProviderMockRecorder) GetMetadataRecord(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMetadataRecord", reflect.TypeOf((*MockProvider)(nil).GetMetadataRecord), arg0, arg1)
 }
