@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: MethodLineRepo)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_app_method_line -destination test/mocks/app/method_line/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces MethodLineRepo
+//
 // Package mock_app_method_line is a generated GoMock package.
 package mock_app_method_line
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	interfaces "github.com/checkmarxDev/ast-sast-export/internal/app/interfaces"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockMethodLineRepo is a mock of MethodLineRepo interface.
@@ -44,7 +48,7 @@ func (m *MockMethodLineRepo) GetMethodLines(arg0, arg1, arg2 string) ([]string, 
 }
 
 // GetMethodLines indicates an expected call of GetMethodLines.
-func (mr *MockMethodLineRepoMockRecorder) GetMethodLines(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockMethodLineRepoMockRecorder) GetMethodLines(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMethodLines", reflect.TypeOf((*MockMethodLineRepo)(nil).GetMethodLines), arg0, arg1, arg2)
 }
@@ -59,7 +63,7 @@ func (m *MockMethodLineRepo) GetMethodLinesByPath(arg0, arg1 string) ([]*interfa
 }
 
 // GetMethodLinesByPath indicates an expected call of GetMethodLinesByPath.
-func (mr *MockMethodLineRepoMockRecorder) GetMethodLinesByPath(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockMethodLineRepoMockRecorder) GetMethodLinesByPath(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetMethodLinesByPath", reflect.TypeOf((*MockMethodLineRepo)(nil).GetMethodLinesByPath), arg0, arg1)
 }
