@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: PresetProvider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_preset_interfaces -destination test/mocks/app/preset/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces PresetProvider
+//
 // Package mock_preset_interfaces is a generated GoMock package.
 package mock_preset_interfaces
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	soap "github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockPresetProvider is a mock of PresetProvider interface.
@@ -44,7 +48,7 @@ func (m *MockPresetProvider) GetPresetDetails(arg0 int) (*soap.GetPresetDetailsR
 }
 
 // GetPresetDetails indicates an expected call of GetPresetDetails.
-func (mr *MockPresetProviderMockRecorder) GetPresetDetails(arg0 interface{}) *gomock.Call {
+func (mr *MockPresetProviderMockRecorder) GetPresetDetails(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetPresetDetails", reflect.TypeOf((*MockPresetProvider)(nil).GetPresetDetails), arg0)
 }
