@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: SourceFileRepo)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_app_source_file -destination test/mocks/app/source_file/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces SourceFileRepo
+//
 // Package mock_app_source_file is a generated GoMock package.
 package mock_app_source_file
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	interfaces "github.com/checkmarxDev/ast-sast-export/internal/app/interfaces"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockSourceFileRepo is a mock of SourceFileRepo interface.
@@ -43,7 +47,7 @@ func (m *MockSourceFileRepo) DownloadSourceFiles(arg0 string, arg1 []interfaces.
 }
 
 // DownloadSourceFiles indicates an expected call of DownloadSourceFiles.
-func (mr *MockSourceFileRepoMockRecorder) DownloadSourceFiles(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockSourceFileRepoMockRecorder) DownloadSourceFiles(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DownloadSourceFiles", reflect.TypeOf((*MockSourceFileRepo)(nil).DownloadSourceFiles), arg0, arg1)
 }
