@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/integration/rest (interfaces: Client)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_integration_rest -destination test/mocks/integration/rest/mock_client.go github.com/checkmarxDev/ast-sast-export/internal/integration/rest Client
+//
 // Package mock_integration_rest is a generated GoMock package.
 package mock_integration_rest
 
@@ -9,7 +13,7 @@ import (
 	reflect "reflect"
 
 	rest "github.com/checkmarxDev/ast-sast-export/internal/integration/rest"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockClient is a mock of Client interface.
@@ -44,7 +48,7 @@ func (m *MockClient) Authenticate(arg0, arg1 string) error {
 }
 
 // Authenticate indicates an expected call of Authenticate.
-func (mr *MockClientMockRecorder) Authenticate(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Authenticate(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Authenticate", reflect.TypeOf((*MockClient)(nil).Authenticate), arg0, arg1)
 }
@@ -59,7 +63,7 @@ func (m *MockClient) CreateScanReport(arg0 int, arg1 string, arg2 rest.Retry) ([
 }
 
 // CreateScanReport indicates an expected call of CreateScanReport.
-func (mr *MockClientMockRecorder) CreateScanReport(arg0, arg1, arg2 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) CreateScanReport(arg0, arg1, arg2 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateScanReport", reflect.TypeOf((*MockClient)(nil).CreateScanReport), arg0, arg1, arg2)
 }
@@ -149,7 +153,7 @@ func (m *MockClient) GetProjects(arg0, arg1, arg2 string, arg3, arg4 int) ([]*re
 }
 
 // GetProjects indicates an expected call of GetProjects.
-func (mr *MockClientMockRecorder) GetProjects(arg0, arg1, arg2, arg3, arg4 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) GetProjects(arg0, arg1, arg2, arg3, arg4 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetProjects", reflect.TypeOf((*MockClient)(nil).GetProjects), arg0, arg1, arg2, arg3, arg4)
 }
@@ -164,7 +168,7 @@ func (m *MockClient) GetProjectsWithLastScanID(arg0, arg1, arg2 string, arg3, ar
 }
 
 // GetProjectsWithLastScanID indicates an expected call of GetProjectsWithLastScanID.
-func (mr *MockClientMockRecorder) GetProjectsWithLastScanID(arg0, arg1, arg2, arg3, arg4 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) GetProjectsWithLastScanID(arg0, arg1, arg2, arg3, arg4 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetProjectsWithLastScanID", reflect.TypeOf((*MockClient)(nil).GetProjectsWithLastScanID), arg0, arg1, arg2, arg3, arg4)
 }
@@ -254,7 +258,7 @@ func (m *MockClient) GetTriagedResultsByScanID(arg0 int) (*[]rest.TriagedScanRes
 }
 
 // GetTriagedResultsByScanID indicates an expected call of GetTriagedResultsByScanID.
-func (mr *MockClientMockRecorder) GetTriagedResultsByScanID(arg0 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) GetTriagedResultsByScanID(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTriagedResultsByScanID", reflect.TypeOf((*MockClient)(nil).GetTriagedResultsByScanID), arg0)
 }
@@ -284,7 +288,7 @@ func (m *MockClient) PostResponseBody(arg0 string, arg1 io.Reader) ([]byte, erro
 }
 
 // PostResponseBody indicates an expected call of PostResponseBody.
-func (mr *MockClientMockRecorder) PostResponseBody(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) PostResponseBody(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PostResponseBody", reflect.TypeOf((*MockClient)(nil).PostResponseBody), arg0, arg1)
 }
