@@ -1,13 +1,17 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/integration/similarity (interfaces: IDProvider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_integration_similarity -destination test/mocks/integration/similarity/provider_mock.go github.com/checkmarxDev/ast-sast-export/internal/integration/similarity IDProvider
+//
 // Package mock_integration_similarity is a generated GoMock package.
 package mock_integration_similarity
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockIDProvider is a mock of IDProvider interface.
@@ -43,7 +47,7 @@ func (m *MockIDProvider) Calculate(arg0, arg1, arg2, arg3, arg4, arg5, arg6, arg
 }
 
 // Calculate indicates an expected call of Calculate.
-func (mr *MockIDProviderMockRecorder) Calculate(arg0, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10 interface{}) *gomock.Call {
+func (mr *MockIDProviderMockRecorder) Calculate(arg0, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Calculate", reflect.TypeOf((*MockIDProvider)(nil).Calculate), arg0, arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)
 }
