@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/integration/soap (interfaces: Adapter)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_integration_soap -destination test/mocks/integration/soap/mock_adapter.go github.com/checkmarxDev/ast-sast-export/internal/integration/soap Adapter
+//
 // Package mock_integration_soap is a generated GoMock package.
 package mock_integration_soap
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	soap "github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockAdapter is a mock of Adapter interface.
@@ -59,7 +63,7 @@ func (m *MockAdapter) GetPresetDetails(arg0 int) (*soap.GetPresetDetailsResponse
 }
 
 // GetPresetDetails indicates an expected call of GetPresetDetails.
-func (mr *MockAdapterMockRecorder) GetPresetDetails(arg0 interface{}) *gomock.Call {
+func (mr *MockAdapterMockRecorder) GetPresetDetails(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetPresetDetails", reflect.TypeOf((*MockAdapter)(nil).GetPresetDetails), arg0)
 }
@@ -89,7 +93,7 @@ func (m *MockAdapter) GetResultPathsForQuery(arg0, arg1 string) (*soap.GetResult
 }
 
 // GetResultPathsForQuery indicates an expected call of GetResultPathsForQuery.
-func (mr *MockAdapterMockRecorder) GetResultPathsForQuery(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAdapterMockRecorder) GetResultPathsForQuery(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetResultPathsForQuery", reflect.TypeOf((*MockAdapter)(nil).GetResultPathsForQuery), arg0, arg1)
 }
@@ -104,7 +108,7 @@ func (m *MockAdapter) GetSourcesByScanID(arg0 string, arg1 []string) (*soap.GetS
 }
 
 // GetSourcesByScanID indicates an expected call of GetSourcesByScanID.
-func (mr *MockAdapterMockRecorder) GetSourcesByScanID(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockAdapterMockRecorder) GetSourcesByScanID(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetSourcesByScanID", reflect.TypeOf((*MockAdapter)(nil).GetSourcesByScanID), arg0, arg1)
 }
