@@ -6,8 +6,8 @@ import (
 	mock_integration_soap "github.com/checkmarxDev/ast-sast-export/test/mocks/integration/soap"
 
 	"github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
 )
 
 var (
