@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: InstallationProvider)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_installation_interfaces -destination test/mocks/app/installation/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces InstallationProvider
+//
 // Package mock_installation_interfaces is a generated GoMock package.
 package mock_installation_interfaces
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	soap "github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockInstallationProvider is a mock of InstallationProvider interface.
