@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/checkmarxDev/ast-sast-export/internal/app/interfaces (interfaces: QueriesRepo)
-
+//
+// Generated by this command:
+//
+//	mockgen -package mock_interfaces_queries -destination test/mocks/app/queries/mock_provider.go github.com/checkmarxDev/ast-sast-export/internal/app/interfaces QueriesRepo
+//
 // Package mock_interfaces_queries is a generated GoMock package.
 package mock_interfaces_queries
 
@@ -8,7 +12,7 @@ import (
 	reflect "reflect"
 
 	soap "github.com/checkmarxDev/ast-sast-export/internal/integration/soap"
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockQueriesRepo is a mock of QueriesRepo interface.
